rating/requirements.txt
alabaster==0.7.12
argh==0.26.2
asgiref==3.2.3
asn1crypto==1.3.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
colorama==0.4.3
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
djangorestframework==3.11.0
docutils==0.16
entrypoints==0.3
flake8==3.7.9
future==0.18.2
helpdev==0.6.10
idna==2.9
imagesize==1.2.0
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.1
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
keyring==21.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
numpydoc==0.9.2
packaging==20.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.6.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.3
psutil==5.7.0
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyQt5==5.12.3
PyQt5-sip==12.7.1
PyQtWebEngine==5.12.1
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.8
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.3
pyzmq==19.0.0
QDarkStyle==2.8
QtAwesome==0.7.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
rope==0.16.0
Rtree==0.9.3
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.0.1
spyder-kernels==1.8.1
sqlparse==0.3.1
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
ujson==1.35
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
yapf==0.29.0
zipp==3.0.0
